fix type cast

+++ openssl/rsa.go
@@ -197,7 +197,7 @@ func SignRSAPKCS1v15(priv *PrivateKeyRSA, h crypto.Hash, hashed []byte) ([]byte,
 func VerifyRSAPKCS1v15(pub *PublicKeyRSA, h crypto.Hash, hashed, sig []byte) error {
 	if pub.withKey(func(pkey *C.EVP_PKEY) C.int {
 		size := C.go_openssl_EVP_PKEY_get_size(pkey)
-		if C.size_t(len(sig)) < size {
+		if len(sig) < int(size) {
 			return 0
 		}
 		return 1
